% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human27pop_labels}
\alias{human27pop_labels}
\title{Labels of 27 human populations}
\format{Labels of 27 human populations. :
\describe{
   \item{human27pop_labels}{It is a vector of labels of 544 individuals. There are 27 populations.}
  ...
}}
\usage{
human27pop_labels
}
\description{
Labels of 27 human populations
}
\keyword{datasets}
