% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipADMIXTURE.R
\name{ipADMIXTURE}
\alias{ipADMIXTURE}
\title{ipADMIXTURE function}
\usage{
ipADMIXTURE(Qmat, admixRatioThs)
}
\arguments{
\item{Qmat}{is a Q matrix that contains admixture ratios of all individuals where the \code{Qmat[i,j]} represents the admixture ratio of ancestor j for individual i.}

\item{admixRatioThs}{is a threshold to determine that if a cluster has \code{meanDiffAdmixRatio} lower than threshold, then the cluster is a homogeneous cluster.}
}
\value{
This function returns sth
}
\description{
ipADMIXTURE is a main function
}
\examples{
# Running ipADMIXTURE on Q matrix of 27 human population dataset where K = 12
h27pop_obj<-ipADMIXTURE(Qmat=human27pop_Qmat, admixRatioThs =0.15)

}
