% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAdmixClusters.R
\name{plotAdmixClusters}
\alias{plotAdmixClusters}
\title{plotAdmixClusters}
\usage{
plotAdmixClusters(obj)
}
\arguments{
\item{obj}{is an object of ipADMIXTURE class.}
}
\description{
plotAdmixClusters is function that plots admixture ratios
where the x axis represents individuals with cluster labels and y axis represents admixture ratios.
}
\examples{
h27pop_obj<-ipADMIXTURE(Qmat=ipADMIXTURE::human27pop_Qmat[[11]], admixRatioThs =0.15)
ipADMIXTURE::plotAdmixClusters(h27pop_obj)

}
