% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClusterLeaves.R
\name{plotClusterLeaves}
\alias{plotClusterLeaves}
\title{plotClusterLeaves}
\usage{
plotClusterLeaves(obj)
}
\arguments{
\item{obj}{is an object of ipADMIXTURE class.}
}
\description{
plotClusterLeaves is function that plots clusters in a form of treemap plot.
Subsquares represent clusters. Each subsquare contains cluster label (ID),
number of members (N), and a maximum of manitude-difference of admixture ratios (md).
A size of each subsquare represents a ratio of member numbers compared to other clusters.
A color represents an md value of cluster.
}
\examples{
h27pop_obj<-ipADMIXTURE(Qmat=ipADMIXTURE::human27pop_Qmat[[11]], admixRatioThs =0.15)
ipADMIXTURE::plotClusterLeaves(h27pop_obj)

}
