% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printClustersFromLabels.R
\name{printClustersFromLabels}
\alias{printClustersFromLabels}
\title{printClustersFromLabels}
\usage{
printClustersFromLabels(obj, labels)
}
\arguments{
\item{obj}{is an object of ipADMIXTURE class.}

\item{labels}{is a vector of labels of all individuals.}
}
\description{
printClustersFromLabels is function that reports that clustering results in text mode.
}
\examples{
h27pop_obj<-ipADMIXTURE(Qmat=ipADMIXTURE::human27pop_Qmat[[11]], admixRatioThs =0.15)
ipADMIXTURE::printClustersFromLabels(h27pop_obj,ipADMIXTURE::human27pop_labels)

}
